requirements.txt
Brotli==1.0.9
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cycler==0.11.0
dash==2.5.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
docopt==0.6.2
et-xmlfile==1.1.0
Flask==2.1.2
Flask-Compress==1.12
fonttools==4.33.3
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.23.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
pipreqs==0.4.11
plotly==5.8.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
six==1.16.0
tenacity==8.0.1
urllib3==1.26.9
Werkzeug==2.1.2
wordcloud==1.8.1
yarg==0.1.9

